+++ backends/hip/ceed-hip-qfunction.c
@@ -177,13 +177,12 @@ int CeedQFunctionCreate_Hip(CeedQFunction qf) {
   // Read source
   char *source;
   ierr = CeedQFunctionGetSourcePath(qf, &source); CeedChkBackend(ierr);
-  const char *funname = strrchr(source, ':') + 1;
-  data->qFunctionName = (char *)funname;
-  const int filenamelen = funname - source;
-  char filename[filenamelen];
-  memcpy(filename, source, filenamelen - 1);
-  filename[filenamelen - 1] = '\0';
-  ierr = CeedHipLoadQFunction(qf, filename); CeedChkBackend(ierr);
+  // Empty source path indicates user must supply Q-Function
+  if (source[0] != '\0') {
+    ierr = CeedQFunctionGetKernelName(qf, &data->qFunctionName);
+    CeedChkBackend(ierr);
+    ierr = CeedHipLoadQFunction(qf, source); CeedChkBackend(ierr);
+  }
 
   // Register backend functions
   ierr = CeedSetBackendFunction(ceed, "QFunction", qf, "Apply",